ingestion/requirements.txt
feedparser==6.0.11
httpx==0.27.2
openai==1.51.0
PyJWT==2.9.0
python-slugify==8.0.4
